Separated modules into a public/private API.
