+++ leb/defcon/_layers.py
@@ -1,14 +1,11 @@
 # © All rights reserved. ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
-# Switzerland, Laboratory of Experimental Biophysics
+# Switzerland, Laboratory of Experimental Biophysics, 2018
 # See the LICENSE.txt file for more details.
 
-from keras.layers import (Conv2D,
-                          Conv2DTranspose,
-                          Conv3D,
-                          MaxPooling2D)
+from keras.layers import Conv2D, Conv2DTranspose, Conv3D
 
-#%% convReLU
-def convReLU(filters, kernel=(3,3), strides=(1,1), **kwargs):
+
+def convReLU(filters, kernel=(3, 3), strides=(1, 1), **kwargs):
     conv = Conv2D(filters, kernel_size=kernel, strides=strides,
                    padding='same',
                    activation='relu',
@@ -16,7 +13,8 @@ def convReLU(filters, kernel=(3,3), strides=(1,1), **kwargs):
                    **kwargs)
     return conv
 
-def deconvReLU(filters, kernel=(3,3), strides=(2,2), **kwargs):
+
+def deconvReLU(filters, kernel=(3, 3), strides=(2, 2), **kwargs):
     deconv = Conv2DTranspose(filters,
                              kernel_size=kernel,
                              strides=strides,
@@ -26,7 +24,8 @@ def deconvReLU(filters, kernel=(3,3), strides=(2,2), **kwargs):
                              **kwargs)
     return deconv
 
-def convReLU_3D(filters, kernel=(3,3,3), strides=(1,1,3), **kwargs):
+
+def convReLU_3D(filters, kernel=(3, 3, 3), strides=(1, 1, 3), **kwargs):
     conv_3D = Conv3D(filters, kernel_size=kernel, strides=strides,
                    padding='same',
                    kernel_initializer='orthogonal',
